fix: correctly handle the default `::fetch` style for statements
We've overridden the default value to our magic -123, which normally works,
but if someone passes in `PDO::FETCH_DEFAULT` manually the method can't
recognize it and throw an unimplemented error. This just makes sure the default
is interpretted the same as -123, better matching the standard.

Added an appropriate test as well.